better merge
