+++ grovedb/src/lib.rs
@@ -42,6 +42,8 @@ pub use element::Element;
 #[cfg(feature = "full")]
 pub use element::ElementFlags;
 #[cfg(feature = "full")]
+pub use merk::proofs::query::query_item::QueryItem;
+#[cfg(feature = "full")]
 use merk::{
     self,
     tree::{combine_hash, value_hash},
@@ -55,7 +57,7 @@ pub use merk::{
         },
         worst_case_costs::WorstCaseLayerInformation,
     },
-    proofs::{query::QueryItem, Query},
+    proofs::Query,
 };
 #[cfg(any(feature = "full", feature = "verify"))]
 pub use query::{PathQuery, SizedQuery};
